p2/requirements.txt
asgiref==3.5.2
Django==4.0.5
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
django-cors-headers==3.12.0
django-environ==0.8.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
pandas